Relocate and refactor schema information classes and tests
Moved `InformationSchema` related classes to a dedicated `infoschema` package for better organization and code clarity. Adjusted import statements and package names accordingly. Minor refactorings were also made in `Schema` and related test files to enhance code readability.
