+++ debezium-server-cratedb/src/main/java/io/debezium/server/cratedb/infoschema/InformationSchemaColumnInfo.java
@@ -3,7 +3,7 @@
  *
  * Licensed under the Apache Software License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
  */
-package io.debezium.server.cratedb;
+package io.debezium.server.cratedb.infoschema;
 
 public record InformationSchemaColumnInfo(
 //        String udtSchema,
