+++ debezium-server-cratedb/src/main/java/io/debezium/server/cratedb/schema/Schema.java
@@ -30,10 +30,9 @@ public record Schema() {
 
     public record Array(I innerType) implements I {
 
-        public static Array of(I i) {
-            return new Array(i);
-        }
-    }
+    public static Array of(I i) {
+        return new Array(i);
+    }}
 
     public record Bit(Number size) implements I {
 
@@ -41,15 +40,15 @@ public record Schema() {
 
     public record Dict(Map<Object, I> fields) implements I {
 
-        public static Dict of() {
+    public static Dict of() {
             return of(new LinkedHashMap<>());
         }
 
-        public static Dict of(Map<Object, I> fields) {
+    public static Dict of(Map<Object, I> fields) {
             return new Dict(fields);
         }
 
-        public static Dict of(Object... keyValues) {
+    public static Dict of(Object... keyValues) {
             var fields = new LinkedHashMap<Object, I>();
             for (int i = 0; i < keyValues.length; i += 2) {
                 fields.put(keyValues[i], (I) keyValues[i + 1]);
@@ -60,7 +59,7 @@ public record Schema() {
 
     public record Coli(Set<I> set) implements I {
 
-        public static Coli of(Set<I> set) {
+    public static Coli of(Set<I> set) {
             // flatMap elements of type Coli
             LinkedHashSet<I> newSet = set.stream().flatMap(e -> {
                 if (e instanceof Coli c) {
@@ -72,18 +71,18 @@ public record Schema() {
             return new Coli(newSet);
         }
 
-        public static Coli of(I... list) {
+    public static Coli of(I... list) {
             var set = new LinkedHashSet<I>(Arrays.stream(list).toList());
             return of(set);
         }
 
-        public static I of(Set<I> setA, Set<I> setB) {
+    public static I of(Set<I> setA, Set<I> setB) {
             var set = new LinkedHashSet<>(setA);
             set.addAll(setB);
             return of(set);
         }
 
-        public static I of(Set<I> setA, I b) {
+    public static I of(Set<I> setA, I b) {
             var set = new LinkedHashSet<>(setA);
             set.add(b);
             return of(set);
