+++ debezium-server-cratedb/src/test/java/io/debezium/server/cratedb/ColumnTypeTest.java
@@ -19,6 +19,8 @@ import org.junit.jupiter.api.Test;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 
+import io.debezium.server.cratedb.infoschema.InformationSchemaColumnDetails;
+import io.debezium.server.cratedb.infoschema.InformationSchemaColumnInfo;
 import io.debezium.server.cratedb.types.ArrayType;
 import io.debezium.server.cratedb.types.BigIntType;
 import io.debezium.server.cratedb.types.ColumnType;
