+++ debezium-server-cratedb/src/test/java/io/debezium/server/cratedb/CrateExperiments.java
@@ -39,6 +39,7 @@ import org.slf4j.LoggerFactory;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 
+import io.debezium.server.cratedb.infoschema.InformationSchemaLoader;
 import io.quarkus.test.common.QuarkusTestResource;
 import io.quarkus.test.junit.QuarkusTest;
