+++ debezium-server-cratedb/src/test/java/io/debezium/server/cratedb/infoschema/InformationSchemaLoaderTest.java
@@ -3,7 +3,7 @@
  *
  * Licensed under the Apache Software License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
  */
-package io.debezium.server.cratedb;
+package io.debezium.server.cratedb.infoschema;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 import static org.junit.Assert.assertEquals;
@@ -15,6 +15,9 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
 
+import io.debezium.server.cratedb.ColumnName;
+import io.debezium.server.cratedb.ColumnTypeManager;
+import io.debezium.server.cratedb.CrateTestResourceLifecycleManager;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
