+++ debezium-server-cratedb/src/test/java/io/debezium/server/cratedb/schema/CrateSQLTest.java
@@ -5,11 +5,11 @@
  */
 package io.debezium.server.cratedb.schema;
 
-import org.junit.jupiter.api.Test;
+import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 
 import java.util.List;
 
-import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
+import org.junit.jupiter.api.Test;
 
 class CrateSQLTest {
     @Test
@@ -18,14 +18,11 @@ class CrateSQLTest {
         var afterSchema = Schema.Dict.of(
                 "id", Schema.Primitive.TEXT,
                 "doc", Schema.Dict.of(
-                        "some-list", Schema.Array.of(Schema.Array.of(Schema.Primitive.BOOLEAN))
-                )
-        );
+                        "some-list", Schema.Array.of(Schema.Array.of(Schema.Primitive.BOOLEAN))));
 
         var statements = CrateSQL.toSQL("test_table", beforeSchema, afterSchema);
         assertThat(statements).isEqualTo(List.of(
-                "ALTER TABLE test_table ADD COLUMN doc['some-list'] ARRAY(ARRAY(BOOLEAN))"
-        ));
+                "ALTER TABLE test_table ADD COLUMN doc['some-list'] ARRAY(ARRAY(BOOLEAN))"));
 
         var statements2 = CrateSQL.toSQL("test_table", afterSchema, afterSchema);
         assertThat(statements2).isEqualTo(List.of());
